Sysinfo_reporter
Sysinfo reporter is now in a place I am happy to leave it. the only
thing that is missing is temp info but for now the project can move on
without it.

some base work was also done on the tcp client built into the reporter
and I am going to explore a python tcp server for testing to verify all
the data is coming in correctly.

This will be the last commit on reporters for a while, I want to Mainly
focus on getting a basic server and a client up in running to have a
minimal viable product to show off.