Adds handshaking and started on db work
Add a computer class to keep track of the current state of the clients
computer there is still some work that needs to be to reflect the change
in the db

adds handshake to get a baseline of the computer for the db and grab the
computer id

adds simple config that just has keys and values. the format is "key
value".

next steps
   - finish db design
   - implement the db
   - update the db when changes happen