Improve the functionality of RMS_TERMINAL
now it is able to get status printing out how many clients are
connected, get a list of all the clients id and name, and request
sys-info from a client (the sscanf doesn't seem to be working quite
right here tho).

Also fixed it so if a client exits it no longer hands, by creating a
cleanup job called before clients vector is used if needed. I am not
super concerned about performance here because clients shouldn't be
connecting and disconnection in mass amounts so performance effect is
limited.

Next I'll focus on the db (assuming I dont get distracted again) as well
as fixing up the reporter to use pthread_sigmask as well as cleaning up
the code a little. after the db is complete we move on to our first CLI
client!!!