Started Refactoring Reporter + display computer on term
refactored the reporter to include everything inside of
rms_reporter_client

there is still a bug that exits incorreclty on ctrl+c probably because
I am still using signals.

added the ability to ingest all computer data in the loggerIngestor
and display it in the termial. next is to generalize the code so it
works for all ingestors, and create the db, after some cleanup of the
reporter