Added grabbing computer from id
can now grab the computer by id on launch if it is already saved to the
db. next is to update it based on any changed values. I should try
and limit the amount of requests that go out, so there should be one for

client version change

os version change

cpu change

system_name change

and if more then 3 of these are set at anygiven time just update
everything so there is just 1 request instead of 3

After that next is to get hdd and networks added to db
and usage info.

after that I might expand the sensors for the system or I'll work on
the cli client. I kinda wanna do http but I am not sure. maybe i'll make
a flask http server to interact with the sqlitedb